requirements.txt
simple_lama_inpainting==0.1.2
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
auth0-python==4.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
datadog-api-client==2.12.0
Deprecated==1.2.13
diffusers==0.14.0
dill==0.3.5.1
distlib==0.3.6
entrypoints==0.4
exceptiongroup==1.1.1
fal-serverless==0.6.29
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==1.1.4
Flask-Cors==3.0.10
flaskwebgui==0.3.5
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gradio==3.28.1
gradio_client==0.1.4
grpc-interceptor==0.15.1
grpcio==1.54.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.0.1
iniconfig==2.0.0
isolate==0.12.0
isolate-proto==0.0.27
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.2
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.9.3.0
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-contrib-python==4.7.0.72
opencv-python==4.9.0.80
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
orjson==3.8.11
packaging==23.1
pandas==2.0.1
piexif==1.1.3
Pillow==9.3.0
platformdirs==3.5.0
pluggy==1.0.0
portalocker==2.7.0
protobuf==4.22.3
psutil==5.9.5
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
regex==2023.5.4
requests==2.28.1
rich==13.3.5
safetensors==0.3.1
sahi==0.11.13
scikit-image==0.19.3
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==1.21.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.26.1
structlog==22.3.0
sympy==1.11.1
tblib==1.7.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.4.12
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.4.0
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.1.0
urllib3==1.26.13
uvicorn==0.22.0
virtualenv==20.23.0
websockets==11.0.2
Werkzeug==1.0.1
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.15.0
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
